service_b/requirements.txt
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
python-dotenv==1.2.2
pika==1.3.2


